:green_heart: Quick fix for new YouTube UI
The new YouTube UI uses Polymer and completely different events than
the spf events from the previous version.

The comments section is dynamic. It's not loaded until you scroll down
to the comments section. Which means we have to wait for the section
to exist before we can inject.

Tried listening to various events but the one that seems to work best is
'yt-register-action'. The problem with this event is that it's generic
and is used for many different actions, so we have to check what the
event fired for.

Extra checks are needed to keep backwards compatibility, as the new UI
could be a phased rollout, and because you can revert back to the old UI
via your settings menu.