:bug: Fix position of button when 'Show more' hidden
Videos that do not have enough data in their info panel (e.g. short
description) will not have a 'Show more' button. In those situations we
need to left-align the 'Show comments' button with the content in the
info panel. But when the 'Show more' button is visible, we need to add
extra left-margin to differentiate between the two buttons.

Signed-off-by: Katherine Bowler <katherine.bowler@gmail.com>