Replace build scripts with Makefile
This feels a bit cleaner than the build scripts.

With Make, you can clearly see all the build rules in one file
(though it can also be split into multiple files if needed).

It's declarative, which makes it more readable (to me) because
you tell Make what you want to achieve, and it deduces how to do
that based on the rules defined.

It's also more efficient, since it won't perform a rebuild step if any
of the dependent files for that rule hasn't changed.

It also gives you parallel execution out of the box with the --jobs
option, otherwise it will execute recipes in serial by default.

It's easy to manage dependencies with Make since you can define
highly composable rules, meaning you can reuse and combine them as
modular components.

Note that I'm using GNU Make on my machine.

For terminology such as rules, recipes, and dependencies, see
https://www.gnu.org/software/make/manual/html_node/Rule-Syntax.html