Add minification to build steps
The `compress` target is not documented because it's not intended for
public use. Only used internally as dependency of `bundle` target.
That is, the user should only have to run `make bundle`.

The minify directory is an order-only prerequisite of the `compress`
target [1] so that the _build dir is only created when needed.

1: https://www.gnu.org/savannah-checkouts/gnu/make/manual/html_node/Prerequisite-Types.html