:construction_worker: Add phony targets for build steps
A phony target is one that is not the name of a file, but just a name
for a recipe to be executed by request.

Phony targets are used to avoid a conflict with a file of the same name,
and to improve performance (because the recipe would be executed every
time since the rule would never create the target file when run).