:fire: Remove obsolete Makefile
A new build solution was introduced in https://github.com/dideler/toggle-youtube-comments/pull/37.

The build now uses NodeJS and npm (node package manager), so we no
longer need Make.

Note this introduces a minor regression, specifically version bumping
which was supported with the Makefile. But that will be added soon as
part of https://github.com/dideler/toggle-youtube-comments/issues/36.