Replace globby with lighter fast-glob
The globby package is built on top of fast-glob, but we're not using
any of the extra features from globby right now so fast-glob is
sufficient for our needs.

Note that fast-glob now supports many of the features that only globby
did at first, such as promises, multiple patterns, negated patterns,
directory expansion, and ignored directories.