:memo: Replace npm author/contributor with team in README
This extension will never be published as an npm package (because that's
not how Chrome extensions are distributed). So there's no benefit to
hiding the author and contributor info inside the package.json since it
will never be used and it's not very discoverable in there.

By creating a Team section in the README, it's more obvious who is
behind the extension. It also removes the unnecessary detail of
differentiating between author and contributor.