:lock: Update more packages reported by npm audit
These packages are updated as a result of `npm audit fix --force`.
The previous commit was from `npm audit fix`, without the force.

This is mainly updating the major version of imagemin-pngquant, which
is a safe change as the update is mainly dropping support for Node v4,
making Node v6 the minimum required version by the package.

https://github.com/imagemin/imagemin-pngquant/compare/bded797...be923c2